+++ src/main/tipc/_instance.ts
@@ -1,5 +1,0 @@
-import { tipc } from '@egoist/tipc/main'
-
-export const t = tipc.create()
-
-export type T = typeof t