+++ src/renderer/src/components/modules/app/WindowsTitlebar.tsx
@@ -1,7 +1,7 @@
 import { isPlayingAtom } from '@renderer/atoms/player'
 import { useWindowState, WindowState } from '@renderer/atoms/window'
 import { ElECTRON_CUSTOM_TITLEBAR_HEIGHT, ELECTRON_WINDOWS_RADIUS } from '@renderer/constants'
-import { tipcClient } from '@renderer/lib/client'
+import { ipcClient } from '@renderer/lib/client'
 import { useAtomValue } from 'jotai'
 
 export const Titlebar = () => {
@@ -26,7 +26,7 @@ export const Titlebar = () => {
         className="no-drag-region hover:bg-muted pointer-events-auto flex h-full w-[50px] items-center justify-center duration-200"
         type="button"
         onClick={() => {
-          tipcClient?.windowAction({ action: 'minimize' })
+          ipcClient?.app.windowAction({ action: 'minimize' })
         }}
       >
         <i className="icon-[mingcute--minimize-line]" />
@@ -36,7 +36,7 @@ export const Titlebar = () => {
         type="button"
         className="no-drag-region hover:bg-muted pointer-events-auto flex h-full w-[50px] items-center justify-center duration-200"
         onClick={async () => {
-          await tipcClient?.windowAction({ action: 'maximum' })
+          await ipcClient?.app.windowAction({ action: 'maximum' })
         }}
       >
         {windowState === WindowState.MAXIMIZED ? (
@@ -50,7 +50,7 @@ export const Titlebar = () => {
         type="button"
         className="no-drag-region pointer-events-auto flex h-full w-[50px] items-center justify-center duration-200 hover:bg-red-500 hover:!text-white"
         onClick={() => {
-          tipcClient?.windowAction({ action: 'close' })
+          ipcClient?.app.windowAction({ action: 'close' })
         }}
       >
         <i className="icon-[mingcute--close-line]" />